Added Scene Transitioning

+++ Components/Structures/GScene.go
@@ -3,7 +3,7 @@ package Structures
 //Scene Configuration
 //Maps to JSON
 type GScene struct {
-	Frames     int
+	Frames     float64
 	Objects    []GObject
 	Animations []*GAnimation
 }
